requirements.txt
networkx
pandas
scikit-learn
openai
anthropic
PyYAML
sentence-transformers
rouge-score
pytest
jupyter
matplotlib
tqdm
numpy
